fix(): type depending on arch

+++ include/db_types.h
@@ -1,11 +1,11 @@
 #ifndef __DB_TYPES_H_
 #define __DB_TYPES_H_ 
 
-// Check 64-bit flags first
+/*Check 64-bit flags first*/
 #if defined(_M_X64) || defined(__x86_64__) || defined(__aarch64__) || defined(_WIN64)
     #define CPU_WORD_SIZE 8
 
-// check 32-bit flags
+/* check 32-bit flags*/
 #elif defined(_M_IX86) || defined(__i386__) || defined(__arm__) || defined(_WIN32)
     #define CPU_WORD_SIZE 4
 #endif
@@ -19,7 +19,7 @@ typedef 	unsigned int	ui32;
 typedef 	int				i32;
 typedef 	unsigned long long	ui64;
 typedef 	long long		i64;
-typedef		long int		file_offset;
+typedef		long long 		file_offset;
 typedef 	long long		process_id;
 #elif CPU_WORD_SIZE == 8
 typedef 	char			i8;
@@ -30,7 +30,7 @@ typedef 	unsigned int	ui32;
 typedef 	int				i32;
 typedef 	unsigned long	ui64;
 typedef 	long		i64;
-typedef		long long 		file_offset;
+typedef		long int		file_offset;
 typedef 	long long		process_id;
 #endif
